# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EntreLibros API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "paths": {
...
    "/api/community/feed": {
      "get": {
        "summary": "Retrieve the paginated community feed",
        "tags": ["Community"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Page number starting from 0"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20,
              "default": 8
            },
            "description": "Number of items per page (max 20)"
          }
        ],
        "responses": {
          "200": {
            "description": "Community feed items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedItem"
                  }
                }
              }
            }
          },
          "400": {
            "description": "community.errors.invalid_pagination"
          }
        }
      }
    },
    "/api/community/posts/{postType}/{id}/like": {
      "parameters": [
        {
          "name": "postType",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": ["listing", "story"]
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^(story-)?[0-9]+$"
          }
        }
      ],
      "post": {
        "summary": "Toggle a persistent community post like",
        "tags": ["Community"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated like state and total",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityLike"
                }
              }
            }
          },
          "401": {
            "description": "auth.errors.unauthorized"
          },
          "404": {
            "description": "community.social.post_not_found"
          }
        }
      }
    },
    "/api/community/posts/{postType}/{id}/comments": {
      "parameters": [
        {
          "name": "postType",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": ["listing", "story"]
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^(story-)?[0-9]+$"
          }
        }
      ],
      "get": {
        "summary": "List ordered comments for a community post",
        "tags": ["Community"],
        "responses": {
          "200": {
            "description": "Community comments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommunityComment"
                  }
                }
              }
            }
          },
          "404": {
            "description": "community.social.post_not_found"
          }
        }
      },
      "post": {
        "summary": "Create a comment on a community post",
        "tags": ["Community"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["body"],
                "properties": {
                  "body": {
                    "type": "string",
                    "maxLength": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created comment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityComment"
                }
              }
            }
          },
          "401": {
            "description": "auth.errors.unauthorized"
          },
          "404": {
            "description": "community.social.post_not_found"
          },
          "422": {
            "description": "community.social.comment_invalid"
          }
        }
      }
    },
    "/api/community/stories": {
      "post": {
        "summary": "Create a community story",
        "tags": ["Community"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommunityStory"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Persisted community story",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryFeedItem"
                }
              }
            }
          },
          "401": {
            "description": "auth.errors.unauthorized"
          },
          "422": {
            "description": "community.story validation error"
          }
        }
      }
    },
...
    "/api/community/discovery": {
      "get": {
        "summary": "Retrieve personalized community discovery data",
        "tags": ["Community"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Stories, relevant books and personalized suggestions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityDiscovery"
                }
              }
            }
          },
          "401": {
            "description": "auth.errors.unauthorized"
          }
        }
      }
    },
    "/api/community/follows/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 1
          }
        }
      ],
      "post": {
        "summary": "Follow a community member",
        "tags": ["Community"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Follow relationship created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowResponse"
                }
              }
            }
          },
          "401": {
            "description": "auth.errors.unauthorized"
          },
          "404": {
            "description": "community.follow.errors.target_not_found"
          },
          "422": {
            "description": "community.follow.errors.self"
          }
        }
      },
      "delete": {
        "summary": "Stop following a community member",
        "tags": ["Community"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Follow relationship removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowResponse"
                }
              }
            }
          },
          "401": {
            "description": "auth.errors.unauthorized"
          },
          "404": {
            "description": "community.follow.errors.target_not_found"
          },
          "422": {
            "description": "community.follow.errors.self"
          }
        }
      }
    },
    "/api/community/corners/nearby": {
      "get": {
        "summary": "List nearby community corners",
        "tags": ["Community"],
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            },
            "description": "Latitude to centre the search"
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            },
            "description": "Longitude to centre the search"
          },
          {
            "name": "radiusKm",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "description": "Search radius in kilometres"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "description": "Maximum number of corners to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Nearby corners ordered by distance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommunityCornerSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid coordinate parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure listing corners",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/community/corners/map": {
      "get": {
        "summary": "Retrieve map pins for community corners",
        "tags": ["Community"],
        "responses": {
          "200": {
            "description": "Pins ready to render in the mini map",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityCornerMap"
                }
              }
            }
          },
          "500": {
            "description": "Unable to generate the corners map",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/community/corners/{id}": {
      "get": {
        "summary": "Retrieve a public community corner detail",
        "tags": ["Community"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public corner detail with approximate location",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityCornerDetail"
                }
              }
            }
          },
          "404": {
            "description": "Corner not found or not publicly available"
          },
          "500": {
            "description": "Unable to load the corner detail"
          }
        }
      },
      "patch": {
        "summary": "Edit a community corner owned by the current user",
        "tags": ["Community"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCornerPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Corner updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityCornerDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required"
          },
          "404": {
            "description": "Corner not found or not owned by the current user"
          },
          "422": {
            "description": "Validation errors"
          }
        }
      }
    },
    "/api/community/corners": {
      "post": {
        "summary": "Publish a new community corner",
        "tags": ["Community"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishCornerPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Corner created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishCornerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation errors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to publish the corner",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionToken"
      }
    },
...
    "schemas": {
...
      "FeedItem": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/BookFeedItem"
          },
          {
            "$ref": "#/components/schemas/SwapFeedItem"
          },
          {
            "$ref": "#/components/schemas/SaleFeedItem"
          },
          {
            "$ref": "#/components/schemas/SeekingFeedItem"
          },
          {
            "$ref": "#/components/schemas/StoryFeedItem"
          }
        ]
      },
      "CommunityLike": {
        "type": "object",
        "properties": {
          "liked": {
            "type": "boolean"
          },
          "likes": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": ["liked", "likes"]
      },
      "CommunityComment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "author", "avatar", "body", "createdAt"]
      },
...
      "CreateCommunityStory": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          },
          "imageUrl": {
            "type": ["string", "null"]
          },
          "bookListingId": {
            "type": ["string", "null"]
          }
        },
        "anyOf": [
          {
            "required": ["body"]
          },
          {
            "required": ["imageUrl"]
          },
          {
            "required": ["bookListingId"]
          }
        ]
      },
...
      "FollowResponse": {
        "type": "object",
        "properties": {
          "following": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": ["following", "userId"]
      },
...
    }
  },
...
}